pyproject.toml
[project]
name = "etl_pipeline"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2",
    "requests>=2.31",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.6",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pytest>=7.0",
    "pydantic>=2.0",
    "tenacity>=8.2"
]